# region cache: batch find regions by key ranges from cache