--- a/internal/locate/sorted_btree.go
+++ b/internal/locate/sorted_btree.go
@@ -92,7 +92,7 @@ func (s *SortedRegions) AscendGreaterOrEqual(startKey, endKey []byte, limit int)
 		if !region.checkRegionCacheTTL(now) {
 			return false
 		}
-		if !region.Contains(lastStartKey) { // uncached hole
+		if len(lastStartKey) > 0 && !region.Contains(lastStartKey) { // uncached hole
 			return false
 		}
 		lastStartKey = region.EndKey()